test fixes, *massive* decrease in log data by eliding large strings
We do this by walking all of the TEST_ONLY_* arguments and scanning for
strings (up to 3 levels deep) that are "long" and replacing them with
strings that say how long they are.  The general assumption is that we
do not need to show snippets because this only happens for gibberish
encrypted data.  In the event we start moving large message payloads
we care about/want to name, it might make sense to replace with a
structured object (by convention) that includes a hash of the text or
the like so we can trace the flow of those huge strings, etc.

For example, the same-proc-group-messaging test was generating a 6.5MB
JSON blob but now it generates 615KB.