main app should now actually create a rawclient instance; but problems.
This commit should now find us spinning up a rawclient instance in the
'client daemon' hidden frame, and find us with the ability for everybody
to talk amongst themselves.  This is untested.

The most immediate and obvious problem is that our crypto keys appear to
be undergoing some type of corruption likely related to string encoding
issues, although it's not clear right now where that would even be
happening.  Unfortunately, the stack on the exception where we are
choking and dying on stops at the compartment boundary, losing us
basically all context.