trying to get the server list displaying, need loggest hookup
It is extremely likely that an exception is being thrown by the
backside while processing the queryServers command.; ideally we
would like a loggest hookup of some type, but anything that ensures
that we see all our exceptions would be delightful.

Although in-browser loggest is slightly over the top in terms of
dependencies, I think it'll be worth it.  It probably wants its
own tab that is accessed hard-core.