status commit: moda conversation creation testing progress
This commit marks initial changes to support moda testhelper's
conversation creation operating by reusing the (refactored) logic in the
server testhelper.  Since it depends on a lot of information at
creation time, we interpose a testwrapper and add the ability to have
the testwrapper tell us the return value of the released function
invocation.  We conveniently also have the backside logic return stuff
that really has no reason to be returned.  (At least we are consistent,
as that is what we do in the RawClient case too :)

same-proc-group-messaging is still passing with this change (after an
unrelated fix from my previous notification status commit.)

The plan is to quickly make similar changes for conversation invitations
and replies.