refactored conv invite and reply test logic for moda's benefit
The moda testhelper stuff is not completely filled out yet, though.  We
are missing the notifications from the rawclient testhelper which needs
to generate livequery expectations.  We then also need to make sure we
can extract a convblurb from somewhere so we can use it to issue our
testhelper commands.