moda helpers generating good expectations, need real moda hookup
The moda test helpers are generating the appropriate peep conversation
query update expectations, but moda is not actually generating the
expectation.  This is expected because I intentionally switched to
writing the unit test half during the implementation to help clarify the
appropriate time to process the updates.

The question was whether it is better to defer most of the notification
generation until the end of the update phase or to perform it as the
events come in.  This is discussed in more details in recent status
commit comments.

The concluded answer is that since peep conv queries are anistropic
(they vary based on the peep being queried on for non "any" "by"
values), it makes sense to perform the iterative logic rather than
attempting to accumulate a big blob of data and create the right
conclusions at the end.  Although we may end up generating annoying
"dancing splices" if a large batch occurs, it is trivial for a post-pass
to perform a compaction to eliminate redundant splice operations.

Additionally, if there gets to be way too much data in a batch (because
of our failure to compact), we can just ship it out early and/or
considering "dropping" the queries' current states and reconstituting
them with a refreshed set of data from the database.

Overall, this should make things much more straightforward for debugging
and logging/event generation because actions should have an obvious and
immediate effect rather than building an ugly complex intermediary state
blob which is harder to log and generally to understand.