status commit: notification generation for conversations, etc.
The end-of-session question is whether we should:

a) add a batch maximization function that fixes up the tuples with the
results of the maximization function, then generate the notifications
off of those, transforming it into an update representation, and thus
letting us need a specific max path.

b) have all the notifications have a specific 'max' case that would
perform the max on the current in-memory value and the new value,
without having to wait for the database result query.

c) similarly (and as a deeper change), store all the index values in the
database so they are always available all the time.

'c' is discarded immediately for being too deep a change at this point
in time, and being the type of thing that is better done with some data
available and something we can optimize.

My call right now is to go with 'a' since it reduces code complexity.
Although one might think it would have an affect on the latency of
notifications, the reality is that we hold back the notifications until
we get the replica-up-to-date notification, which waits on the all of
the tasks completing which in turn wait on the I/O ops to complete, so
there is absolutely no affect on latency at this time.

If we can get storage implementations that can issue blind 'max'
operations by virtue of coprocessor/fancy storage or we start storing
all the values in memory, we should reconsider then.