support an unordered log/test match mode, moda-shakedown convmsgs works!
The test support for the query mechanism is much improved now because we
don't need to worry about the order the results show up in.  We are
going to need to introduce an additional dep-induces-query-update call
or two on the test side as/if things get a little more complex.

The query test logic is not verifying all the data matches up correctly
after the fact on its own.  We will probably want to move the
conversation blurb query checks to be automatic things in their own
post-update test stage.  That will take care of the dep value checks.