about:loggest-server: use names from about:loggest logic for sanity
So, the server is not generating any names in its log, so we steal them
from the logic we use in about:loggest.  Unfortunately, this still does
not cover all of the crypto keys that can be seen.  Specifically, the
client keys are a problem, since only the clients and the server know
those.  I will probably end up surfacing them directly from the server
at some point because the debug functionality is well isolated and it
would be stupidly useful.

This also cleans up the about:loggest-server setup a little bit more
by eliminating a setTimeout from the equation.