have state rep actually track ordering; fix many ordering bugs
The clever annotation rep might have been written with the assumption
that we would perform comparison based on the attribute ordering, but
the testing logic did not check it.  For simplicity (and slightly out of
fear about heuristics that turn objects into hashmaps), we set the value
of the state objects to their count since we weren't using it for
anything anyways and it does the right thing.

As a related change, we no longer generate preAnno/postAnno values for
moves which we had been setting to 0 previously.  The main reason for
doing this is that moves can be noisy in cases where our incremental
processing may shuffle things back into the same position they started
from.  A performance concern is that this can generate silly splice
shuffles we are going to serialize, but the previously theorized
optimization post-pass should be able to handle that by building a
displacement map representation if we really want to clean up our
output.  I doubt it is likely to save much CPU after we factor in the
analysis pass unless our lists are really huge...

A bunch of typos/brain glitches were fixed in the comparators, and our
comparators are now arbitrarily stable by sorting on the root key for
peeps when the index comparator finds equal values, which is ugly but
super-stable.

There is still some test unhappiness in regards of an extra conversation
query notification, about to address that.