ui test: signup reliably occurring, race conditions dealt with
We should now have a mechanism in place so that the waitForModa stuff
does not have problems racing the client daemon.  The risk is that it's
somewhat granular; in many cases, we are just waiting on 'query' and if
it ends up that there is a lot of stuff going on, we could end up with a
ton of queries all firing around the same time.