test ui: shows potential friends, tab assertion failure
The test is running further every other time, tab assertion failure;
snafu with the invariant expectations I think for when we are already on
home.

The 'every other time' bit is annoying... our signup success callback
event does not appear to be making it across the selenium bridge to its
server.  This is odd because my helpful LOGGY global variable shows that
we are doing the right thing in the Firefox process.  An extenuating
factor could be that I think every other time I am also forgetting to
close the Firefox window before running the test, so that could be
having fallout, though how that would impact anything is questionable.
On the other hand, every other time I ran selenium's own unit tests, it
would screw up every other time for me way early...