status commit; moda bridge data rep almost all happy again
It looks like the simplified namespaceItemModified logic may be going
the wrong way in one of the delta-notification-only cases.  The other
main fallout I expect to need to address is that we no longer will
trigger direct query complete notifications for dependencies, so those
either need to stop being generated (easiest), or need to slave off of
per-item on() invocations.