mob ui: group conversation mods work, but potential race on query/delta
In testing I just observed an event series:
- query X sent with references to A and B but not containing A, B
- query Y sent with A and B

The problem is that _fillOutQueryDepsAndSend splices the needed items
out, so if another _fillOutQueryDepsAndSend occurs during the time after
the splice and before the fulfillment, then it will have a corrupt
result.

A complication that will influence the fix is the potential problem of
delaying queries that don't need data-fills until all pending queries
are completed.  Unfortunately, it doesn't seem like there is a simple
solution unless we implement a dirty-generation concept with reverse
dependencies so we can climb the dependency chain to propagate the
generation data (and then scan each query's items before letting it be
eligible for dispatch.)  That's pretty complicated and I am thinking
ditch it.

Probaby easier to just try and have long-running queries be smart about
their batching so they maintain a minimal dataNeeded overhead at any
given time.  Actually, yeah, exactly that.