convpub read watermark status commit, query suppression fix, testfixes
Most all of the convpub metadata non-UI work should be there now, we
just need to get us some unit tests.

But while prepping for that, I noticed some non-determinism in the
moda-shakedown tests.  The ordering of B and C when their timestamps
are identical is based on the lexicographic order of their root public
keys.  Since that's random, we were seeing success differences depending
on the relation of the keys.  Because of transitions to and from
time-based comparisons and root key comparisons, this might or might not
result in shifts in location and might or might not be mooted by
suppression.

In any event it was really confusing and we were not clearing the
pending bit on suppression, which would break something down the road,
which was confusing.  I also enhanced the moda test logic to understand
how to do the suppression logic itself, which I think may have been the
other variation in failures I was observing.