newness tracking: moda-shakedown test is green with constraints
The moda testhelper framework is now capable of generating expectations
in regards to newness tracking.  This includes both new messages
received inducing newness and marking messages read automatically
clearing newness.  moda-shakedown constructs a newness query alongside
its other conversation queries before conversation activity starts.

Inspection of the test logs shows that the right expectations are being
generated and fulfilled.  The important caveat is that we are not
generating object representations for newness when we generate a new
conversation activity summary for a conversation, just on modifications.
This is a deficiency that happens for everything, but is usually
remedied by an explicit content check (which also helps sanity check the
expectation logic...)

Next up is the introduction of a "moda-newness" test that exercsies the
explicit 'clear newness without marking read' functionality and our
persistence logic for newness.