work on Firefox 11+; MozWebSocket is now WebSocket
Note that it looks like we won't work on Firefox 10.  I think the JS
engine may have had a bug in that release, because we were fine in 9,
but busted in 10, and fine again in 11.  I also recall davida having
things not work on a nightly long enough ago that it might have been
10's nightly...