be compatible with jetpack trunk
The rewritten module loader no longer supported our AMD syntax for our
main.js.  We had no great reason for using the fully expanded AMD
syntax, so it's no biggie.

nacl.js needed a require statement on its own line, presumably because
of a brittle scanner under the new regime?

We need to explicitly request to be unpacked via package.json because
our our binary libraries that jsctypes needs to be able to get at on the
filesystem.